version 1.2 剩下JointTrain的model

+++ options/base_options.py
@@ -1,4 +1,5 @@
 import argparse
+import copy
 import logging
 import os
 
@@ -9,7 +10,7 @@ import losses
 import models
 import networks
 import task_datasets
-from util.util import rmdirs, get_log_level, seed_everything, load_best_ckptname
+from util.util import rmdirs, get_log_level, seed_everything, load_best_ckptname, Checker
 
 
 class BaseOptions(object):
@@ -36,15 +37,12 @@ class BaseOptions(object):
         parser.add_argument('--word_size', type=int, default=1, help='the number of machine or process')
 
         # logging configuration
-        parser.add_argument('--log_filename', type=str, default="output/{}.txt", help='logging filename')
-        # parser.add_argument('--log_filename', type=str, default=None, help='logging filename')
+
         parser.add_argument('--log_filemode', type=str, default='a', help='logging filemode')
         parser.add_argument('--log_format', type=str, default='%(asctime)s - %(levelname)s - %(message)s',
                             help='logging format')
         parser.add_argument('--log_level', type=str, default="debug", help='logging level')
-        # training before
-        parser.add_argument('--dels', type=str, default="",
-                            help='which need to be clear delete, strings contains ckpt, log or output ')
+
         # basic parameters
         parser.add_argument('--name', type=str, default='',
                             help='name of the experiment. opt.name=opt.model_name+"_"+opt.dataset_list')
@@ -62,6 +60,9 @@ class BaseOptions(object):
                             help='customized suffix: opt.name = opt.name + suffix: e.g., {opt.name}_{opt.suffix}')
         parser.add_argument('--preffix', default='', type=str,
                             help='customized preffix: opt.log_filename = opt.preffix+opt.name: e.g., {opt.log_filename}={opt.preffix}_{opt.name}')
+
+        # for setting inputs
+        parser.add_argument('--imsize', type=int, default=256)
 
         self.initialized = True
         return parser
@@ -141,6 +142,9 @@ class BaseOptions(object):
         opt.log_filename = opt.log_filename.strip().lower()
 
         # initialize num_classes, get info from dataset_list, and it will be used in networks
+
+        checker = Checker(dataset_list=opt.dataset_list)
+
         opt.num_classes = task_datasets.get_num_classes_by_data_list(opt.dataset_list)
 
         opt.isTrain = self.isTrain  # train or test
@@ -152,7 +156,7 @@ class BaseOptions(object):
         # log_filename
         if opt.log_filename != 'none':
             opt.log_filename = opt.log_filename.format(f'{opt.preffix}_{opt.name}')
-            if "output" in opt.dels and os.path.isfile(opt.log_filename):
+            if hasattr(opt, "dels") and "output" in opt.dels and os.path.isfile(opt.log_filename):
                 os.remove(opt.log_filename)
             os.makedirs(os.path.dirname(opt.log_filename), exist_ok=True)
         else:
@@ -169,16 +173,18 @@ class BaseOptions(object):
         logging.info(f'Name={opt.name:*^50}')
 
         # clear(delete) the directories
-        if "log" in opt.dels and os.path.isdir(opt.logs_dir):
+        if hasattr(opt, "dels") and "log" in opt.dels and os.path.isdir(opt.logs_dir):
             rmdirs(opt.logs_dir)
 
-        if "ckpt" in opt.dels and os.path.isdir(opt.checkpoints_dir):
+        if hasattr(opt, "dels") and "ckpt" in opt.dels and os.path.isdir(opt.checkpoints_dir):
             rmdirs(opt.checkpoints_dir)
 
         os.makedirs(opt.checkpoints_dir, exist_ok=True)
         os.makedirs(opt.logs_dir, exist_ok=True)
+        os.makedirs(opt.result_dir, exist_ok=True)
 
         load_taskindex, load_epoch = load_best_ckptname(opt.checkpoints_dir)
+
         if opt.load_taskindex == 0 and load_taskindex is not None:
             opt.load_taskindex = load_taskindex
         if opt.load_epoch == 'best' and load_epoch is not None:
@@ -187,6 +193,7 @@ class BaseOptions(object):
         self.print_options(opt)
 
         # set GPUs
+        gpu_ids_str = copy.copy(opt.gpu_ids)
         str_ids = opt.gpu_ids.split(',')
         opt.gpu_ids = []
         for str_id in str_ids:
@@ -195,11 +202,14 @@ class BaseOptions(object):
                 opt.gpu_ids.append(id)
 
         # set gpu ids
-        # os.environ["CUDA_VISIBLE_DEVICES"] = opt.gpu_ids #当指定这个的时候，比如gpu_ids=[1,2], 那系统会把1当做0,2当做1，则在DataParallel中的device_ids就是[0,1]才可以，而不能是[1,2]
+        os.environ[
+            "CUDA_VISIBLE_DEVICES"] = gpu_ids_str  # 当指定这个的时候，比如gpu_ids=[1,2], 那系统会把1当做0,2当做1，则在DataParallel中的device_ids就是[0,1]才可以，而不能是[1,2]
         if min(opt.gpu_ids) >= 0:
-            torch.cuda.set_device(min(opt.gpu_ids))  # 这样就不会出现那种问题
+            # torch.cuda.set_device(min(opt.gpu_ids))  # 这样就不会出现那种问题
+            # opt.device = torch.device('cuda:{}'.format(opt.gpu_ids[0])) if opt.gpu_ids else torch.device(
+            #     'cpu')  # get device tag: CPU or GPU
 
-        opt.device = torch.device('cuda:{}'.format(opt.gpu_ids[0])) if opt.gpu_ids else torch.device(
-            'cpu')  # get device tag: CPU or GPU
+            opt.device = torch.device('cuda:0')
+
         self.opt = opt
         return self.opt
